% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_data.R
\name{rescale_data}
\alias{rescale_data}
\title{Rescale numeric data by standardising or normalising as appropriate}
\usage{
rescale_data(df, scale = c("standardise", "normalise"))
}
\arguments{
\item{df}{a data.frame object}

\item{scale}{Choose type of recaling, default is to 'standardise'
so rescale using the standard deviation. choosing 'normalise'
will scale the data in range [0,1] and therefore outliers will
be less relavnt.}
}
\value{
A data frame where numeric data has been rescaled
}
\description{
Rescale numeric data by standardising or normalising as appropriate
}
\examples{
rescale_data(df = mtcars, scale = "standardise")
}
